Fix security issue in caching script checker
